+++ CHANGES.md
@@ -1,4 +1,6 @@
 
+  - improved documentation: applicative data type / complexity
+
 # 1.0.0
   - first opam package